+++ server/tests-hspec/Test/LongIdentifiersSpec.hs
@@ -22,7 +22,7 @@ import Prelude
 -- Preamble
 
 spec :: SpecWith TestEnvironment
-spec =
+spec = do
   Context.run
     [ -- Create table fails currently becasuse we postfix table names for some reason
       -- which makes the valid table name go over the limit
@@ -72,15 +72,26 @@ spec =
     ]
     tests
 
+  Context.run
+    [ Context.Context
+        { name = Context.Backend Context.Postgres,
+          mkLocalTestEnvironment = Context.noLocalTestEnvironment,
+          setup = Postgres.setup schema,
+          teardown = Postgres.teardown schema,
+          customOptions = Nothing
+        }
+    ]
+    testInsert
+
 --------------------------------------------------------------------------------
 -- Schema
 
 schema :: [Schema.Table]
-schema = [regular, longtable]
+schema = [regular, longtable, longtable2, multitable]
 
 regular :: Schema.Table
 regular =
-  (table "regular")
+  (table "regular_table_with_a_long_name_to_test_rename_identifiers")
     { tableColumns =
         [ Schema.column "id" Schema.TInt,
           Schema.column "name" Schema.TStr
@@ -105,13 +116,55 @@ longtable =
       tableReferences =
         [ Schema.Reference
             { referenceLocalColumn = "regular_id",
-              referenceTargetTable = "regular",
+              referenceTargetTable = "regular_table_with_a_long_name_to_test_rename_identifiers",
               referenceTargetColumn = "id"
             }
         ],
       tableData =
         [ [Schema.VInt 1, Schema.VInt 1, Schema.VInt 1, Schema.VInt 1],
           [Schema.VInt 2, Schema.VInt 2, Schema.VInt 2, Schema.VInt 2]
+        ]
+    }
+
+longtable2 :: Schema.Table
+longtable2 =
+  (table "i_need_a_table_with_a_long_name_to_test_rename_identifiers2") -- 59 characters
+    { tableColumns =
+        [ Schema.column "id" Schema.TInt,
+          Schema.column "i_need_a_column_with_a_long_name_to_test_rename_identifiers" Schema.TInt, -- 59 characters
+          Schema.column "i_need_a_column_with_a_long_name_but_is_different" Schema.TInt -- 49 characters
+        ],
+      tablePrimaryKey = ["id"],
+      tableData =
+        [ [Schema.VInt 1, Schema.VInt 1, Schema.VInt 1],
+          [Schema.VInt 2, Schema.VInt 2, Schema.VInt 2]
+        ]
+    }
+
+multitable :: Schema.Table
+multitable =
+  (table "table_with_multiple_relationships_and_long_name_2_test_rewrite") -- 62 characters
+    { tableColumns =
+        [ Schema.column "id" Schema.TInt,
+          Schema.column "reference_for_longtable_id" Schema.TInt,
+          Schema.column "reference_for_longtable2_id" Schema.TInt
+        ],
+      tablePrimaryKey = ["id"],
+      tableReferences =
+        [ Schema.Reference
+            { referenceLocalColumn = "reference_for_longtable_id",
+              referenceTargetTable = "i_need_a_table_with_a_long_name_to_test_rename_identifiers",
+              referenceTargetColumn = "id"
+            },
+          Schema.Reference
+            { referenceLocalColumn = "reference_for_longtable2_id",
+              referenceTargetTable = "i_need_a_table_with_a_long_name_to_test_rename_identifiers2",
+              referenceTargetColumn = "id"
+            }
+        ],
+      tableData =
+        [ [Schema.VInt 1, Schema.VInt 1, Schema.VInt 1],
+          [Schema.VInt 2, Schema.VInt 2, Schema.VInt 2]
         ]
     }
 
@@ -173,7 +226,7 @@ data:
           testEnvironment
           [yaml|
 query {
-  hasura_regular(order_by:[{id:asc}]) {
+  hasura_regular_table_with_a_long_name_to_test_rename_identifiers(order_by:[{id:asc}]) {
     id
     i_need_a_table_with_a_long_name_to_test_rename_identifierss_by_id_to_regular_id(order_by:[{i_need_a_column_with_a_long_name_to_test_rename_identifiers:asc, i_need_a_column_with_a_long_name_but_is_different:asc}]) {
       i_need_a_column_with_a_long_name_to_test_rename_identifiers
@@ -185,7 +238,7 @@ query {
       )
       [yaml|
 data:
-  hasura_regular:
+  hasura_regular_table_with_a_long_name_to_test_rename_identifiers:
   - id: 1
     i_need_a_table_with_a_long_name_to_test_rename_identifierss_by_id_to_regular_id:
       - i_need_a_column_with_a_long_name_to_test_rename_identifiers: 1
@@ -195,3 +248,132 @@ data:
       - i_need_a_column_with_a_long_name_to_test_rename_identifiers: 2
         i_need_a_column_with_a_long_name_but_is_different: 2
 |]
+
+testInsert :: Context.Options -> SpecWith TestEnvironment
+testInsert opts = do
+  it "insert to regular table" $ \testEnvironment -> do
+    shouldReturnYaml
+      opts
+      ( GraphqlEngine.postGraphqlYaml
+          testEnvironment
+          [yaml|
+          query: |
+            mutation MyMutation(
+                $x: Int
+              ) {
+              insert_hasura_regular_table_with_a_long_name_to_test_rename_identifiers
+                ( objects:
+                  { name: "bbb",
+                    id: $x
+                  }
+                ) {
+                affected_rows
+                returning {
+                  id
+                  name
+                }
+              }
+            }
+          variables:
+            x: 3
+          |]
+      )
+      [yaml|
+      data:
+        insert_hasura_regular_table_with_a_long_name_to_test_rename_identifiers:
+          affected_rows: 1
+          returning:
+          - id: 3
+            name: bbb
+      |]
+
+  it "nested insert to long table name" $ \testEnvironment -> do
+    shouldReturnYaml
+      opts
+      ( GraphqlEngine.postGraphqlYaml
+          testEnvironment
+          [yaml|
+          query: |
+            mutation MyMutation(
+                $x: Int
+              ) {
+              insert_hasura_regular_table_with_a_long_name_to_test_rename_identifiers
+                ( objects:
+                  { name: "bbb",
+                    id: 4,
+                    i_need_a_table_with_a_long_name_to_test_rename_identifierss_by_id_to_regular_id:
+                      { data:
+                        { i_need_a_column_with_a_long_name_but_is_different: 4,
+                          i_need_a_column_with_a_long_name_to_test_rename_identifiers: $x,
+                          id: 4
+                        }
+                      }
+                  }
+                ) {
+                affected_rows
+                returning {
+                  id
+                  name
+                }
+              }
+            }
+          variables:
+            x: 4
+          |]
+      )
+      [yaml|
+      data:
+        insert_hasura_regular_table_with_a_long_name_to_test_rename_identifiers:
+          affected_rows: 2
+          returning:
+          - id: 4
+            name: bbb
+      |]
+
+  it "nested insert to table with multiple relationships" $ \testEnvironment -> do
+    shouldReturnYaml
+      opts
+      ( GraphqlEngine.postGraphqlYaml
+          testEnvironment
+          [yaml|
+          query: |
+            mutation MyMutation(
+                $x: Int
+              ) {
+              insert_hasura_table_with_multiple_relationships_and_long_name_2_test_rewrite
+                ( objects:
+                  { id: 5,
+                    i_need_a_table_with_a_long_name_to_test_rename_identifiers_by_reference_for_longtable_id_to_id:
+                      { data:
+                        { i_need_a_column_with_a_long_name_but_is_different: 5,
+                          i_need_a_column_with_a_long_name_to_test_rename_identifiers: $x,
+                          id: 5,
+                          regular_id: 1,
+                        }
+                      },
+                    i_need_a_table_with_a_long_name_to_test_rename_identifiers2_by_reference_for_longtable2_id_to_id:
+                      { data:
+                        { i_need_a_column_with_a_long_name_but_is_different: 5,
+                          i_need_a_column_with_a_long_name_to_test_rename_identifiers: $x,
+                          id: 5
+                        }
+                      }
+                  }
+                ) {
+                affected_rows
+                returning {
+                  id
+                }
+              }
+            }
+          variables:
+            x: 5
+          |]
+      )
+      [yaml|
+      data:
+        insert_hasura_table_with_multiple_relationships_and_long_name_2_test_rewrite:
+          affected_rows: 3
+          returning:
+          - id: 5
+      |]